additional functionality needed for LibHealpix.jl
I am currently in the processing of rewriting a lot of LibHealpix.jl.
One of its limitations has been that it uses the libchealpix functions
to read and write Healpix maps as FITS images. These functions only work
with Float32 values, so we were frequently discarding precision when
doing I/O with Healpix maps.

To remedy that situation I am replacing this part of the wrapper with
the FITSIO Libcfitsio interface. This makes it easy for me to port the
original C functions into Julia.

This commit adds some missing functionality that I needed to make this
work:

* wrap `fits_read_key_str`
* wrap `fits_read_key_lng`
* wrap `fits_read_keys_lng`
* wrap `fits_write_date`
* widen the signature of `fits_write_key` to accept all `Number`s

ref issue JuliaAstro/FITSIO.jl#61